pyproject.toml
[tool.poetry]
name = "nexusai"
version = "0.1.0"
description = "Enterprise AI Agent Orchestration Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
pydantic = "^2.6.0"
sqlalchemy = "^2.0.25"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
uvicorn = "^0.27.0"
asyncio = "^3.4.3"
aiohttp = "^3.9.3"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"
pytest-asyncio = "^0.23.4"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
